Fix Pr failure ticket #273
 Long version:  In August 2013 it was determined that many instability problems experienced by SasView were do to a threading problem in fitting.py.  Mainly the authors seemed to not realize that calling thread.stop just raised a flag and that the thread itself needed to check for said flag regularly and kill the thread if it found the flag raised.  A quick fix was implemented by adding a wait till the existing thread ended naturally to the two places in the code where the stop is raised.  This seems to work.  

At the same time pr.py was found to use the same raising of the stop flag three times and it was assumed that the same problem existed and the same wait “fix” was implemented.  Subsequenty it has been shown that the new wait code in pr.py NEVER gets reached and thus the fix breaks the PrView perspective.  It may be that this earlier code properly dealt with threads properly(still remembering how the thread code was meant to be used).  Whatever the case will remove this fix for now.   

Finally another thread.stop was found once in simulation.py.  since this is older code it may also be handled properly so am removing the “Fix” from that as well for now.

to quickly find these points in the future: 
fitting.py - two instances that seem to need fixing
pr.py - three instances that may work correctly
simulation.py - one instace

search terms:
“-PDB” or
“d.stop” (thread.stop or thread_1D.stop or 2D.stop)