Final fixes (hopefully) to Batch Grid
1) Added back right click on label to allow insertion of columns
2) Added back left click on label to allow correct handling of set axix
for plots
3) Added override for sheet.CSheet.PaintBackground which has not been
updated to account for the need to always have a dc object on a paint
event in wx.Python 3 -- probably the main problem with sheet - edit was
working but generating a pile of errors behind the scenes.
4) Fixed the exception raised on insert column submenu due to removing
Edit menu. Needed to remove chunk of code or grab the wx.Menu handle.
Did later.
5) Added lots of doc strings but most are not rendering so need to
figure out why at some point - but not a release issue.