Merge master
see issue #590. fix error in q.rg and uncertainties in Rg and I0.
Uncertainty calculation for slope and intercept are still wrong I
believe.

references #205 - Removed Kratky plot choice from linearized plots for
now since it doesn't currently work.

close #590.  A lot of problems were caused by LineModel and LinerFit
having different equations (ax+b vs a+bx).  Further errors in
calculations, particularly of uncertainties were fixed. The fact that
the fits to not account for smearing was verified and a warning added.
Code was also modified to update the qmin and qmax to match changes in
the transformed xmin xmax.  Lots of documentation was added and the
fitdialog layout was cleaned up considerably.  This is now usable though
the design of the user interface (and the whole design of linear fits)
could use a rethink.