Improve efficiency of CanSAS 2.0 reader
When it came to reading the Q and I data values, the reader was looping
through every single data point in the hdf5 file and appending it to an
array. This caused problems when loading very large files, rendering
SasView unresponsive. Now, the reader saves the entire array from the
file, instead of looping through each point.