Handle missing columns in .ses format
The uncertainties on spin echo length and lambda aren't strictly
required, so we should handle files that do not have them.  SasView
seems to be unhappy with an uncertainty of zero, so we've put in a
small default value.  If the user isn't happy with the default value,
then they should have measured the real one.