Fix viewer documentation
Replaced the three images for the image_viewer_help.rst with new ones
that reflect the latest features. Also replaced the bmp with png.
Finaly added a comment in the image_viewer.py code noting the close
error generated by PIL (for any non png images). This should complete
the work on this branch.