Problem when copying multiplicy models caused crash
crashed when multiplicy model was selected eg spherical sld, because there was a drop down box instead of high/low like it was expecting.
there are is now error handling for this but it does not copy and paste the drop down box selection which will need to be implimented seperately.
When GUI loaded Paste is now disabled until user copies params. Paste also checks if model has changed since last copy, and displays message warning user that not all params may paste, with option to cancel. Also - focus fixed such that if paste is clicked the params automatically update without having to move mouse to reset the focus.