Fix reporting of pinhole resolution on FitPage when taken from data
GUI now reports dQ/Q as the label suggested.  Also labels were fixed to
be more clear.  Finally a small bit of code cleanup was attempted.  Note
that there is a serious problem with the 2D resolution part of that GUI.
This is a previously unreported problem which may be dealt with in
another ticket or may be done here depending on the effort required.

addresses #1206