change default to fold=false and prep for arbitrary boxslice.
Default GUI had fold=True forcing -qi to be averaged with +qi which in
principle gives better statistics IF symmetric.  The general case is not
to average.  Moreover the average (in sascalc) seems to be forced to
fold=False so not clear that in fact we are currently even getting the
advantage of statistics!!

Also started process of allowing the GUI to use any xmin, xmax; ymin,
ymas box, whether or not it includes q=0,0.