adding model 1D data sets
Adding 2 data sets created using 5.0.1 theory with an added header
detailing exactly the parameters used and giving the calculation for Sv
and the porod constant.  Initially this was for the unit tests of the
invariant but also added to 1D test files folder since very few if any
have enough information currently to know if one is obtaining the right
parameters when trying to fit them.