First unit test Added with dev docs
First in a number of commits to fix the invariant unit tests.  Will be
replacing 2 sets of tests with 3 new ones using well characterized data.
The first test of the first series along with the class docstrings
detailing the data and the calculation of the correct values along with
caveats.