Fix Sv calculation, error and display
The Sv was being calculated in a circular manner and provided completely
bogus uncertainty.  Sv is fundamentally based on the porod constant and
contrast with are both user inputs with NO uncertainty estimate (would
be nice to get some?).  Redid the computation to use those values
directly so that it will display even if vol fraction has an error.
Also replaced error with NONE if ds ==0. I actually included code do
properly compute the uncertainty if uncertainties are ever given on
Porod Constant and contrast..

Also replace NaN with ERROR  both for s and phi and added a logger
statement that goes to the log file but currently NOTHING is going to
the console screen as mentioned in another ticket somewhere.