Fix vol fraction uncertainty calculation
The vol frac uncertainty calculation uses dy of course. but if the data
had not dy, the code decided to create a dy=sqrt(y).  This is only valid
for raw counts not reduced data. Creating a completely bogus uncertainty
in order to then propagate is just wrong.  Changed so that if the data
has no uncertainty the vol frac uncertainty is given as -1

Actually the value is bogus anyway since the real uncertainty depends on
how much of the curve is being captured correctly but that we can't do
much about.