Updated AnnulusSlicer to improve slicer min limits (low-q 2D data) and max limits (offset 2D data)
Changed the evaluation of min() to max() value to allow slicing of offset 2D data without limiting to the qx value closer to offset center. Outer radius can now be typed in the SlicerParameters window, limited by the edge farthest from the offset center.

Defined the minimum annulus difference based on the 2d data set, rather than some arbitrary small number, in which the minimum slice will depend on the minimum q in the dataset.