Second stage of invariant update port from 4.x
updated the sascalc file now. Again looks like it can be a direct copy
as there seems to have been no changes between master and ESS_GUI.  this
should fix the actual factor of 2 as well as the totally bogus
uncertainties.  Left todo would be the help.rst and a couple of GUI
fixes.