change Sv code based on review comments
The code in the PR was incorret as pointed out by P. Kienzle, since the
value in the uncertainty box is set from a call at the end of the
routine.  Only if the call to surface fails would setting the value
work.

Instead we leave InvariantPerspective as originally written and make the
changes in invarian.py.  ds now returns None if it is not computed and a
value if it is. InvariantPerspective seems to handle it ok.