Disentangle total Q* from Q* from data
as recorded in PR #1590 the invariant total was first being reported
only as the invariant from data. That was fixed but then the status
dialog with the breakdown of the Q* components reported the total Q*
instead of that from data.  This is because the entire invariant panels
did not distingish between the two.