Fix problem with vol_fraction calc and tweak to front matter in invariant.py
Added extrapolation to the vol fraction call so that it returns the
value based on the chosen extrapolation not just from Q*data. Also
realized that the problem encountered by Steve of the import from future
not being the first import was due to eclipse being helpful and adding
an import pi which I did not even need. So reverted to the original
(without the pi import and moving the from future import to after the
doc string.