Make new child plots unselected by default
Any perspective creating child data and adding them to the datamanager
now default to them being unselected. If a perspective needs them to be
selected instantly it will need to deliberately set the checkstate at
Checked.  This should normally be avoided as it causes no end of
confusion to the user.