improvement of edge cases in fourier transform
The previous version of this algorithm simply added a small epsilon value to each Q to ensure the vectors did not line up with the element faces - which would cause divide by zero errors - instead special case algorithms have been implemented so epsilon is only used where necessary - and L'Hospital is used where possible.

This is the first working case of this algorithm and several improvements have yet to be added, such as direction specific epsilon vectors, and exploring the range for which L'Hospital's limit can be used.