fixed volume correction
the re-ordering of setting parameters and data in the GUI (required for theta and phi to be transformed) prevented volume correction working, due to the re-calculation in set_sld_data. This has been made optional with a default as true (to make scripting logical - volume should not NEED to be explicitely set as a parameter to get hte correct answer) and the GUI does not apply it. Since the GUI also resets the volume with parameters this is not an issue.