Model encoded as a dictionary
Created a dictionary within the FittingWidget class, named model_dict,  to contain the main models of SasView. The keys are “standard”, “poly” and “magnet” relating to the main model, polydisperse model and magnet model respectively.
Many functions in the code have been refactored to accept an argument, called model_key, which corresponds to a key in model_dict. This allows the model to be passed to a function without having to pass the bumps object, which can lead to the code misbehaving.