Stop using __file__ to find resources
Using __file__ to find resources is hopelessly complicated across the different
frozen modules, debug modules, unpacked source etc that are used to run sasview.
importlib.resource provides a uniform interface to finding resources that are
shipped alongside modules.

Since the `config_file` member is not used anywhere outside this class,
let's also make it 'private' since it's a file-like object that should
be private to the module.