Suppress SIGSEGV errors from pytest
The GUI tests consistently exit with segmentation faults. It is likely
that these are the result of the widgets not being disposed of in the
correct order, meaning that the destructors for the Qt objects are run
twice.

This is a latent bug in the test suite (and perhaps in SasView's Qt GUI)
that needs to be fixed one day.