up default points on slicer and fix a few related slicer bugs
the slicers defaulted to 20 or 36 points (depending on the slicer). This is almost never a reasonable number and there have been several request for that to be made larger. Upped it to 100 here.

Noted in testing that the box slicer validators were not properly checking for 0 points (the other slicers did) . In fact they were do nothing functions so fixed that by adding the appropriate code snipet.

Also all the slicers seemed to have the default set in a couple of different places which caused some odd behavior in at least on case. that was cleaned up as well.