Change duplicate class name
There are three unused slicer modules that are set up for wedge type slicers. These were never implemented. I suspect that in fact it was orinally planed to use this for sector averaging where perhaps the wedge became a sector by making r1=0 and r2 = qmax. However as it stands it left two slicers classes called SectorInteractor.  This unused class is renamed to the more appropriate WedgeInteractor in preparation for being used for crystal type data.