Temporary fix to issue with the two classes called being a bit too different
The purpose of this method is to create and update the 1D plot from the sector selections. The problem is that the phi and Q averaging are different enough that this method contains three instances of If (check class name) do x. The point of post_data vs _post_data I think was to avoid this kind of thing an suggests that in this case we may need a new method in the WedgeInteracgtorPhi and WedgeInteracgtorQ to handle these specifics? But that seems a bit complicated -- leaving as is for now by adding this third "if class name" check to fix the immediate problem.