Taking care of spurious file changes
Changes were made to release 5.0.5 branch and attempted to be reverted 2 years ago. For reasons lost in the mists of time, 3 months later branch 5.0.5 was merged into this branch even though it was based on main and not being released?  This led to 4 files from @Dehoni work in a completely different branch becoming part of the changes in this branch compared to main. By usind checkout main --file, I've now updated the offending files with what is currently in main (the base) so that they should no longer be part of the pull request.