Clean up the width nomenclature and simplify width calc code.
The review comments asked for the dictionary set params for x and y widths be standardized. In fact it turned out that a) the widths were actually half widths and b) there was a variety of words used. The entire BoxSlicer.py code now uses half_width and half_height (and half_width_min etc).

Also as per review, the abs(half_width/height) from set params was removed. It turns out the values should never be negative but it is unclear whether absoluting them if they end up negative (which should not be possible at this point)  will yield the correct answer. Thus better for it to fail.

also corrected the nbins calculation in _post_data for direction="Y" (the plus should have been a minus)