make graphical changes to ROI more robust
ROI has no data error is impossible to catch without checking to see if any of the data are in the ROI. This is effectively done in manipulations.py as part of deciding what to average. Rather than check twice, use a try except to capture the ValueError that manipulations.py raises and report it in the log then negate the move.