One Final bon bon for box Slicer
connect the top and right lines (those with the markers) so they can be moved without "catching" the marker. Particularly useful if the box becomes very narrow so that the center marker and line marker are on top of each other.  This also matches the behavior of the sector slicer so also part of making slicers consistent.