limit log.warning to one for each drag error
Implemented a self.move_valid flag in order to only throw 1 logging.warning message until a valid move occurs to reset the flag.  Unfortuantely this cannot be implemented yet in the onDrag of BaseInteractor as that would change the interface for all other classes deriving from BaseInteractor. Instead have removed the logging for now. In fact this check is no longer necessary for box interactor and not sure any other interactor except boxsum perhaps? So probably eventually remove that check from BaseInteractor.py altogether.

Also changed the logging.error to logging.warning on the try except catch. TODO: either change all to loggin.error, lower the leve at which it starts the console log ... or decide not telling the user is OK.