Change behavior of moving ROI box outside of data
Instead of ignoring moves that move an edge of the ROI box outside the data, reset the edge to the datamin or datamax as appropriate. This makes it easier to move a box to the edge of the data box and also makes it possible to vertically move a horizontal strip going from xmin to xmax etc.