make vol frac plot of size distribution linear logf and add errorbars
Basically we cannout use any ROLES as they always overide any customization. Then, in poltter.py, force linear for that plot in the same way that the Dmax plot was forced to log linear.  For error bars on 'non data'we must use "hide_error=False". For some reason "show_error" only is an attribute of the thype is data. Not clear that we could not make it type data?