Final tweaks to the model editor.
All changes are updating the templates in some way. Lots of editing of the recommenations. Also replace VR whith shell_form as VR is no longer supported. Note that currently the templates include several python functions which should NOT be included for a C model when a C model is being created. The current thought seems to be that since the c model cannot easily be written, we are making sure the file generated will still run. This should probalby be discussed.