Fixes default weighting behavior
checkbox 2 enabling and default checked checkbox logic was moved from enableButtons and setupWindow to setData. Those defaults depend on whether or not the new data set has dy or not and should not be changed everytime I fit or other action is performed (besides loading new data), nor should they be set permanently at the start.

Also fixed a bug in logic where di_flag was not getting reset once set to "True". This could have been done in the main perspective on cleanup between data sets but seemed more general to fix at the source