Fix last problem with reparameterization editor
Problem of editing parameters after creating them was that the tool was doing weird thing if the selected item was one of the items in the parameter list instead of the parameter itself. Particularly for the Description which has its own child.

This is pointless anyway as the full parameter editor is started in any case and the user will change more than what they clicked on most likely.