Update tutorials
Added the new corefunc tutorial from the Tutorials repo and edited the tutorials.rst to replace the v5 with this v6.  Also added the create custom models tutorial to the rst list. The PDF had been added but the second step of adding to the rst file had been missed.

Finally deleted the obsolete PDFs since there are not used. The originals are still in the Tutorials repo