Fix to document download urls not working
Just had to capture the download request signal and accept it. However python files to not seem to actually download. Also our implementatin of QtWebEngine makes the download so silent the use thinks nothing is happening so added a logging.warning to force the log explorer open and a message generated